market별 coin 가격 단위 로직 변경

+++ seonoh/ArchitectureStudy/app/src/main/java/com/example/seonoh/seonohapp/CoinItemViewHolder.kt
@@ -19,7 +19,7 @@ class CoinItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     fun bind(data: UseCoinModel) {
         coinNameTv.text = data.market
         currentPriceTv.text = data.tradePrice
-        changeRateTv.setTradeDiff(data.signedChangeRate.toDouble())
+        changeRateTv.setTradeDiff(data.signedChangeRate)
         totalTradePriceTv.text = data.accTradePrice_24h
 
     }
